+++ system/src/documents/ActorSD.mjs
@@ -835,6 +835,12 @@ export default class ActorSD extends Actor {
 	}
 
 
+	async getPatron() {
+		const uuid = this.system.patron ?? "";
+		return await this._getItemFromUuid(uuid);
+	}
+
+
 	async getDeity() {
 		const uuid = this.system.deity ?? "";
 		return await this._getItemFromUuid(uuid);
@@ -854,6 +860,34 @@ export default class ActorSD extends Actor {
 		}
 
 		return rollData;
+	}
+
+
+	async getSpellcasterClasses() {
+		const actorClass = await this.getClass();
+
+		const playerSpellClasses = [];
+
+		let spellClass = actorClass.system.spellcasting.class;
+		if (spellClass === "") {
+			playerSpellClasses.push(actorClass);
+		}
+		else if (spellClass !== "__not_spellcaster__") {
+			playerSpellClasses.push(
+				await this._getItemFromUuid(spellClass)
+			);
+		}
+
+		const spellcasterClasses =
+			await shadowdark.compendiums.spellcastingBaseClasses();
+
+		for (const bonusClass of this.system.bonuses.spellcastingClasses ?? []) {
+			playerSpellClasses.push(
+				spellcasterClasses.find(c => c.name.slugify() === bonusClass)
+			);
+		}
+
+		return playerSpellClasses.sort((a, b) => a.name.localeCompare(b.name));
 	}
 
 
@@ -907,13 +941,19 @@ export default class ActorSD extends Actor {
 	}
 
 
-	async isSpellcaster() {
+	async isSpellCaster() {
 		const characterClass = await this.getClass();
 
 		const spellcastingClass =
 			characterClass?.system?.spellcasting?.class ?? "__not_spellcaster__";
 
-		return characterClass && spellcastingClass !== "__not_spellcaster__"
+		const isSpellcastingClass =
+			characterClass && spellcastingClass !== "__not_spellcaster__";
+
+		const hasBonusSpellcastingClasses =
+			(this.system.bonuses.spellcastingClasses ?? []).length > 0;
+
+		return isSpellcastingClass || hasBonusSpellcastingClasses
 			? true
 			: false;
 	}
@@ -993,6 +1033,48 @@ export default class ActorSD extends Actor {
 		return gearSlots;
 	}
 
+
+	async openSpellBook() {
+		const playerSpellcasterClasses = await this.getSpellcasterClasses();
+
+		const openChosenSpellbook = classUuid => {
+			new shadowdark.apps.SpellBookSD(
+				classUuid,
+				this.id
+			).render(true);
+		};
+
+		if (playerSpellcasterClasses.length <= 0) {
+			return ui.notifications.error(
+				game.i18n.localize("SHADOWDARK.item.errors.no_spellcasting_classes"),
+				{ permanent: false }
+			);
+		}
+		else if (playerSpellcasterClasses.length === 1) {
+			return openChosenSpellbook(playerSpellcasterClasses[0].uuid);
+		}
+		else {
+			return renderTemplate(
+				"systems/shadowdark/templates/dialog/choose-spellbook.hbs",
+				{classes: playerSpellcasterClasses}
+			).then(html => {
+				const dialog = new Dialog({
+					title: game.i18n.localize("SHADOWDARK.dialog.spellbook.open_which_class.title"),
+					content: html,
+					buttons: {},
+					render: html => {
+						html.find("[data-action='open-class-spellbook']").click(
+							event => {
+								event.preventDefault();
+								openChosenSpellbook(event.currentTarget.dataset.uuid);
+								dialog.close();
+							}
+						);
+					},
+				}).render(true);
+			});
+		}
+	}
 
 	/** @inheritDoc */
 	prepareData() {