+++ system/src/documents/CompendiumsSD.mjs
@@ -203,6 +203,10 @@ export default class CompendiumsSD {
 		return CompendiumsSD._documents("Item", "NPC Features", filterSources);
 	}
 
+	static async patrons(filterSources=true) {
+		return CompendiumsSD._documents("Item", "Patron", filterSources);
+	}
+
 	static async patronBoonTables(filterSources=true) {
 		const documents = await CompendiumsSD._documents(
 			"RollTable", null, filterSources
@@ -275,11 +279,20 @@ export default class CompendiumsSD {
 		return sources.sort((a, b) => a.name.localeCompare(b.name));
 	}
 
+	static async spellcastingBaseClasses(filterSources=true) {
+		const documents = await CompendiumsSD._documents("Item", "Class", filterSources);
+
+		return this._collectionFromArray(documents.filter(document =>
+			document.system.spellcasting.class === ""
+		));
+	}
+
 	static async spellcastingClasses(filterSources=true) {
 		const documents = await CompendiumsSD._documents("Item", "Class", filterSources);
+
 		return this._collectionFromArray(documents.filter(document =>
 			document.system.spellcasting.ability !== ""
-			&& document.system.spellcasting.class !== "NONE"
+			&& document.system.spellcasting.class !== "__not_spellcaster__"
 		));
 	}