+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ## Enhancements
 * [#657] Consume Ammunition When Attacking With Ranged Weapons if Required
+* [#847] Add support for Warlock class Boon which gives the ability to learn a Wizard spell
 * [#901] Added Russian as a partially complete system language.
 * [#920] Show tooltips on weapon and armor properties when showing expanded inline view in inventory
 
@@ -20,6 +21,8 @@
 	- Settings for compendium art mapping are now found in *Settings > Core > Compendium Art*
 
 * [#917] Restructure Item sheet templates to make them a bit more manageable
+* [#927] Migration runner accessing globalThis.duplicate which must now be accessed via foundry.utils.duplicate
+* [#928] Migrate "Deities" to new "Patrons and Deities" compendium
 * [#929] Merged Russian language updates from Crowdin
 
 *Many thanks to **AdamsGH** for contributing **Russian** translation data.*